refactor: updates styles import to use Intervention module
Switches components to import styles from the Intervention stylesheet
for improved consistency and maintainability. Removes the now-unused
PlantLocation styles file to clean up redundant code.
