+++ src/features/projects/components/Intervention/ImageSlider.tsx
@@ -3,7 +3,7 @@ import type { Story } from 'react-insta-stories/dist/interfaces';
 import { useState, useEffect } from 'react';
 import Stories from 'react-insta-stories';
 import getImageUrl from '../../../../utils/getImageURL';
-import styles from './../../styles/PlantLocation.module.scss';
+import styles from './../../styles/Intervention.module.scss';
 
 export type SliderImage = {
   image?: string;
