+++ src/features/projects/components/maps/Interventions.tsx
@@ -11,7 +11,7 @@ import React from 'react';
 import { Layer, Marker } from 'react-map-gl';
 import { Source } from 'react-map-gl';
 import { useProjectProps } from '../../../common/Layout/ProjectPropsContext';
-import styles from '../../styles/PlantLocation.module.scss';
+import styles from '../../styles/Intervention.module.scss';
 import * as turf from '@turf/turf';
 import { localizedAbbreviatedNumber } from '../../../../utils/getFormattedNumber';
 import { useLocale, useTranslations } from 'next-intl';
